ovirt: Interogate clusters and templates for the survey
Motivation
When running the installer on oVirt, you need to decide on which oVirt
cluster to create VMs and which RHCOS template to use. It is much
nicer to pick the cluster from the list, same for the template.

Modification
Make the survey get the clusters and the templates from oVirt

Result
The survery will use the go-ovirt sdk to fetch all needed data from
oVirt and will ask the user to pick a cluster and a template.

Internally the platform object will store the IDs of the cluster and the
template and not the names.

Signed-off-by: Roy Golan <rgolan@redhat.com>